+++ src/adapters/index.ts
@@ -1,0 +1,3 @@
+export * from './data-adapter';
+export * from './empty-adapter';
+export * from './xlsx';
