+++ src/index.ts
@@ -1,4 +1,5 @@
-export * from './populator';
-export * from './adapters/data-adapter';
+export * from './populators/populator';
+export { AbstractDataAdapter, EmptyDataAdapter, XLSXAdapter } from './adapters';
+export * from './models';
 export * from './populators/abstract-populator';
 export * from './utils/create-logger';
