+++ src/populators/populator.ts
@@ -1,14 +1,14 @@
 import { CherrytwistClient } from 'cherrytwist-lib';
 import { Logger } from 'winston';
-import { AbstractDataAdapter } from './adapters/data-adapter';
-import { AbstractPopulator } from './populators/abstract-populator';
-import { ChallengePopulator } from './populators/challenge-populator';
-import { EcoversePopulator } from './populators/ecoverse-populator';
-import { GroupPopulator } from './populators/group-populator';
-import { HostPopulator } from './populators/host-populator';
-import { OpportunityPopulator } from './populators/opportunity-populator';
-import { OrganisationPopulator } from './populators/organisation-populator';
-import { UserPopulator } from './populators/user-populator';
+import { AbstractDataAdapter } from '../adapters/data-adapter';
+import { AbstractPopulator } from './abstract-populator';
+import { ChallengePopulator } from './challenge-populator';
+import { EcoversePopulator } from './ecoverse-populator';
+import { GroupPopulator } from './group-populator';
+import { HostPopulator } from './host-populator';
+import { OpportunityPopulator } from './opportunity-populator';
+import { OrganisationPopulator } from './organisation-populator';
+import { UserPopulator } from './user-populator';
 
 export class Populator extends AbstractPopulator {
   // Create the ecoverse with enough defaults set/ members populated
